Own polynomial.factor in the domain bundle and finish test-CI cleanup
Move univariate factorization into the polynomial DomainBundle as a
ComputedOperation with RationalPolynomial contracts, a shared SymPy kernel,
and a flint exact-replay checker. Retire the legacy PolynomialFactorAdapter
install path and dual artifact/semantics folklore.

Also migrate remaining exact-domain recipes onto open_exact_domain_services,
split the MCP adapter hub by concern, and ratchet architecture checks against
direct installer and complete-runtime fixture imports outside owning confests.
