+++ src/jacobian/contracts/polynomials.py
@@ -262,61 +262,6 @@ class PolynomialKellerConditionVerifyRequest(ContractModel):
         return self
 
 
-class PolynomialFactorRequest(ContractModel):
-    variable: PolynomialVariable
-    polynomial: SparseRationalPolynomial
-
-    @model_validator(mode="after")
-    def require_univariate_terms(self, info: ValidationInfo) -> Self:
-        if any(len(term.exponents) != 1 for term in self.polynomial.terms):
-            raise ValueError("factorization currently supports univariate polynomials")
-        if canonicalization_preflight_active(info):
-            return self
-        if len(self.polynomial.terms) > 512:
-            raise ValueError("factorization exceeds the 512-term operation budget")
-        for term in self.polynomial.terms:
-            require_bounded_rational(
-                term.coefficient,
-                max_digits=256,
-                label="factorization coefficient",
-            )
-            if term.exponents[0] > 127:
-                raise ValueError(
-                    "factorization exceeds the degree-127 operation budget"
-                )
-        return self
-
-
-class PolynomialFactorRecord(ContractModel):
-    factor: SparseRationalPolynomial
-    multiplicity: int = Field(ge=1, le=127)
-
-
-class PolynomialFactorizationArtifact(ContractModel):
-    factorization_schema_version: Literal["1"] = "1"
-    variable: PolynomialVariable
-    source_polynomial_uri: ArtifactUri
-    coefficient: CanonicalRational
-    factors: tuple[PolynomialFactorRecord, ...] = Field(max_length=1024)
-    reconstructed: SparseRationalPolynomial
-    backend: Literal["sympy"] = "sympy"
-    backend_version: str = Field(min_length=1, max_length=64)
-
-
-class PolynomialFactorOutput(ContractModel):
-    source_polynomial_uri: ArtifactUri
-    factorization_uri: ArtifactUri
-    variable: PolynomialVariable
-    coefficient: CanonicalRational
-    factors: tuple[PolynomialFactorRecord, ...]
-    reconstructed: SparseRationalPolynomial
-    exactness: Literal["EXACT"] = "EXACT"
-    product_reconstruction: Literal["EXACT"] = "EXACT"
-    irreducibility_verification: Literal["UNVERIFIED"] = "UNVERIFIED"
-    backend: Literal["sympy"] = "sympy"
-    backend_version: str
-
-
 class PolynomialCollisionRequest(ContractModel):
     first_evaluation_uri: ArtifactUri
     second_evaluation_uri: ArtifactUri
