+++ tests/boundary/providers/flint/runtime/test_linear_rational_solution.py
@@ -3,16 +3,13 @@ from __future__ import annotations
 from pathlib import Path
 
 from tests.support.capabilities import invoke_capability
-from tests.support.services import open_domain_services
+from tests.support.exact_domain import open_exact_domain_services
 
 from jacobian.contracts.capabilities import (
     CapabilityAssuranceLevel,
     CapabilityRequest,
 )
 from jacobian.domains.rational_linear import build_rational_linear_bundle
-from jacobian.exact_domain_checkers import install_exact_domain_verification
-from jacobian.operation_installation import OperationInstaller
-from jacobian.runtime import CheckerAuthorityMode
 
 
 def _system() -> dict[str, object]:
@@ -31,29 +28,10 @@ def _system() -> dict[str, object]:
 
 
 def test_solution_candidate_is_inline_and_replayable(tmp_path: Path) -> None:
-    bundle = build_rational_linear_bundle()
-    with open_domain_services(
-        tmp_path, checker_authority=CheckerAuthorityMode.NONE
+    with open_exact_domain_services(
+        tmp_path,
+        build_rational_linear_bundle(),
     ) as services:
-        installed = OperationInstaller(
-            services.core.store,
-            services.core.schemas,
-            services.core.artifacts,
-        ).install(bundle)
-        for adapter in installed.adapters:
-            services.installation.register_capability(adapter)
-        adapters, _ = install_exact_domain_verification(
-            services.core.store,
-            services.core.schemas,
-            services.core.artifacts,
-            services.installation.verification,
-            services.core.checkers,
-            bundles={"rational_linear": (bundle, installed)},
-            authorize=True,
-        )
-        for adapter in adapters:
-            services.installation.register_capability(adapter)
-
         computed = invoke_capability(
             services, "linear.rational_solution.compute", _system()
         )
