+++ tests/component/providers/matrix/test_linear_rational_inconsistency.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import pytest
 from tests.support.capabilities import invoke_capability
-from tests.support.services import open_domain_services
+from tests.support.exact_domain import open_exact_domain_services
 
 from jacobian.contracts.capabilities import (
     CapabilityAssuranceLevel,
@@ -21,9 +21,6 @@ from jacobian.domains.rational_linear.protocol import (
     parse_inconsistency_worker_response,
     parse_solution_worker_response,
 )
-from jacobian.exact_domain_checkers import install_exact_domain_verification
-from jacobian.operation_installation import OperationInstaller
-from jacobian.runtime import CheckerAuthorityMode
 
 
 def _system() -> dict[str, object]:
@@ -95,29 +92,10 @@ def test_rational_linear_worker_payloads_bind_status_and_source_dimensions() -> 
 
 
 def test_inconsistency_candidate_is_inline_and_replayable(tmp_path: Path) -> None:
-    bundle = build_rational_linear_bundle()
-    with open_domain_services(
-        tmp_path, checker_authority=CheckerAuthorityMode.NONE
+    with open_exact_domain_services(
+        tmp_path,
+        build_rational_linear_bundle(),
     ) as services:
-        installed = OperationInstaller(
-            services.core.store,
-            services.core.schemas,
-            services.core.artifacts,
-        ).install(bundle)
-        for adapter in installed.adapters:
-            services.installation.register_capability(adapter)
-        adapters, _ = install_exact_domain_verification(
-            services.core.store,
-            services.core.schemas,
-            services.core.artifacts,
-            services.installation.verification,
-            services.core.checkers,
-            bundles={"rational_linear": (bundle, installed)},
-            authorize=True,
-        )
-        for adapter in adapters:
-            services.installation.register_capability(adapter)
-
         computed = invoke_capability(
             services,
             "linear.rational_inconsistency.compute",
