+++ tests/component/providers/matrix/test_matrix_hermite_normal_form.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 import pytest
 from tests.support.capabilities import invoke_capability
-from tests.support.services import open_domain_services
+from tests.support.exact_domain import open_exact_domain_services
 
 from jacobian.contracts.capabilities import (
     CapabilityAssuranceLevel,
@@ -14,9 +14,6 @@ from jacobian.contracts.capabilities import (
 from jacobian.contracts.matrices import IntegerMatrix
 from jacobian.domains.matrix_lattice import build_matrix_bundle
 from jacobian.domains.matrix_lattice.hnf import _parse_hnf_worker_result
-from jacobian.exact_domain_checkers import install_exact_domain_verification
-from jacobian.operation_installation import OperationInstaller
-from jacobian.runtime import CheckerAuthorityMode
 
 
 def _matrix(entries: list[list[int]]) -> dict[str, object]:
@@ -31,28 +28,10 @@ def _matrix(entries: list[list[int]]) -> dict[str, object]:
 
 @pytest.fixture
 def hnf_services(tmp_path: Path):
-    bundle = build_matrix_bundle()
-    with open_domain_services(
-        tmp_path, checker_authority=CheckerAuthorityMode.NONE
+    with open_exact_domain_services(
+        tmp_path,
+        build_matrix_bundle(),
     ) as services:
-        installed = OperationInstaller(
-            services.core.store,
-            services.core.schemas,
-            services.core.artifacts,
-        ).install(bundle)
-        for adapter in installed.adapters:
-            services.installation.register_capability(adapter)
-        adapters, _ = install_exact_domain_verification(
-            services.core.store,
-            services.core.schemas,
-            services.core.artifacts,
-            services.installation.verification,
-            services.core.checkers,
-            bundles={"matrix": (bundle, installed)},
-            authorize=True,
-        )
-        for adapter in adapters:
-            services.installation.register_capability(adapter)
         yield services
