+++ tests/domain/matrix/test_legacy_cleanup.py
@@ -14,8 +14,8 @@ from __future__ import annotations
 from pathlib import Path
 
 import pytest
+from tests.support.exact_domain import open_exact_domain_services
 from tests.support.rationals import rational_payload as _q
-from tests.support.services import open_domain_services
 
 from jacobian.contracts.capabilities import (
     CapabilityAssuranceLevel,
@@ -23,8 +23,6 @@ from jacobian.contracts.capabilities import (
 )
 from jacobian.contracts.results import ExecutionStatus
 from jacobian.domains.matrix_lattice import build_matrix_bundle
-from jacobian.exact_domain_checkers import install_exact_domain_verification
-from jacobian.operation_installation import OperationInstaller
 
 
 def _matrix(rows: list[list[int]]) -> dict[str, object]:
@@ -38,31 +36,11 @@ def _matrix(rows: list[list[int]]) -> dict[str, object]:
 @pytest.fixture
 def pilot_matrix_runtime(tmp_path: Path):
     """Open a runtime with only the pilot matrix bundle and its ExactReplay checkers."""
-    with open_domain_services(
+
+    with open_exact_domain_services(
         tmp_path,
         build_matrix_bundle(),
-        checker_authority=__import__(
-            "jacobian.runtime",
-            fromlist=["CheckerAuthorityMode"],
-        ).CheckerAuthorityMode.INSTALL_BUNDLED,
     ) as services:
-        bundle = build_matrix_bundle()
-        installed = OperationInstaller(
-            services.core.store,
-            services.core.schemas,
-            services.core.artifacts,
-        ).install(bundle)
-        adapters, _installation = install_exact_domain_verification(
-            services.core.store,
-            services.core.schemas,
-            services.core.artifacts,
-            services.installation.verification,
-            services.core.checkers,
-            bundles={"matrix": (bundle, installed)},
-            authorize=True,
-        )
-        for adapter in adapters:
-            services.installation.register_capability(adapter)
         yield services
